+++ database/migrations/2018_05_12_071210_create_ts_activities_table.php
@@ -17,11 +17,11 @@ class CreateTsActivitiesTable extends Migration
         Schema::create('ts_activities', function (Blueprint $table) {
             
             // Keys
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->unsignedInteger('client_id');
             $table->unsignedInteger('creator_id');
-            $table->unsignedInteger('schedule_id');
-            $table->unsignedInteger('job_id');
+            $table->unsignedBigInteger('schedule_id');
+            $table->unsignedBigInteger('job_id');
 
             // Attributes
             $table->dateTime('from');
